Move CompiledJarFile class to util class
